+++ sme_terceirizadas/escola/management/commands/atualiza_escolas.py
@@ -7,22 +7,7 @@ from requests import ConnectionError
 
 from ....dados_comuns.constants import DJANGO_EOL_API_TOKEN, DJANGO_EOL_API_URL
 from ...models import Escola
-
-
-def calcula_total_alunos_por_escola(quest_json_data: dict) -> dict:
-    escola_total = {}
-    escolas_quantidades_alunos = quest_json_data['results']
-
-    for escola_data in escolas_quantidades_alunos:
-        codigo_escola = escola_data['cd_escola']
-        total_corrente = escola_data['total']
-        if codigo_escola not in escola_total:
-            escola_total[codigo_escola] = total_corrente
-        else:
-            escola_total[codigo_escola] += total_corrente
-
-    return escola_total
-
+from .helper import calcula_total_alunos_por_escola
 
 env = environ.Env()
